Move both RotationMatrix and Transform classes from drake/multibody/multibody_tree/math to drake/math
