+++ math/BUILD.bazel
@@ -55,6 +55,7 @@ drake_cc_library(
         "random_rotation.cc",
         "roll_pitch_yaw.cc",
         "rotation_matrix.cc",
+        "transform.cc",
     ],
     hdrs = [
         "axis_angle.h",
@@ -65,11 +66,12 @@ drake_cc_library(
         "roll_pitch_yaw_using_quaternion.h",
         "rotation_matrix.h",
         "rotation_matrix_deprecated.h",
+        "transform.h",
     ],
     deps = [
+        "//common:default_scalars",
         "//common:essential",
         "//common:is_approx_equal_abstol",
-        "//multibody/multibody_tree/math:rotation_matrix",
     ],
 )
 
@@ -348,6 +350,13 @@ drake_cc_googletest(
 )
 
 drake_cc_googletest(
+    name = "rotation_matrix_test",
+    deps = [
+        ":geometric_transform",
+    ],
+)
+
+drake_cc_googletest(
     name = "rotation_matrix_deprecated_test",
     deps = [
         ":geometric_transform",
@@ -370,6 +379,13 @@ drake_cc_googletest(
         ":saturate",
         "//common:autodiff",
         "//common:symbolic",
+    ],
+)
+
+drake_cc_googletest(
+    name = "transform_test",
+    deps = [
+        ":geometric_transform",
     ],
 )
