+++ multibody/multibody_tree/BUILD.bazel
@@ -114,7 +114,6 @@ drake_cc_library(
         ":spatial_inertia",
         "//drake/common:autodiff",
         "//drake/math:geometric_transform",
-        "//drake/multibody/multibody_tree/math:rotation_matrix",
     ],
 )
