+++ multibody/multibody_tree/math/BUILD.bazel
@@ -64,46 +64,11 @@ drake_cc_library(
     ],
 )
 
-drake_cc_library(
-    name = "rotation_matrix",
-    srcs = ["rotation_matrix.cc"],
-    hdrs = ["rotation_matrix.h"],
-    deps = [
-        "//drake/common",
-        "//drake/common:default_scalars",
-    ],
-)
-
-drake_cc_library(
-    name = "transform",
-    srcs = ["transform.cc"],
-    hdrs = ["transform.h"],
-    deps = [
-        ":rotation_matrix",
-        "//drake/common",
-        "//drake/common:default_scalars",
-    ],
-)
-
 drake_cc_googletest(
     name = "spatial_algebra_test",
     deps = [
         ":spatial_algebra",
         "//drake/common:autodiff",
-    ],
-)
-
-drake_cc_googletest(
-    name = "rotation_matrix_test",
-    deps = [
-        ":rotation_matrix",
-    ],
-)
-
-drake_cc_googletest(
-    name = "transform_test",
-    deps = [
-        ":transform",
     ],
 )
