+++ multibody/multibody_tree/space_xyz_mobilizer.cc
@@ -7,7 +7,6 @@
 #include "drake/common/eigen_types.h"
 #include "drake/math/roll_pitch_yaw.h"
 #include "drake/math/rotation_matrix.h"
-#include "drake/multibody/multibody_tree/math/rotation_matrix.h"
 #include "drake/multibody/multibody_tree/multibody_tree.h"
 
 namespace drake {
@@ -40,8 +39,8 @@ const SpaceXYZMobilizer<T>& SpaceXYZMobilizer<T>::SetFromRotationMatrix(
   DRAKE_ASSERT(q.size() == kNq);
   // Project matrix to closest orthonormal matrix in case the user provides a
   // rotation matrix with round-off errors.
-  const RotationMatrix<T> Rproj_FM =
-      RotationMatrix<T>::ProjectToRotationMatrix(R_FM);
+  const math::RotationMatrix<T> Rproj_FM =
+      math::RotationMatrix<T>::ProjectToRotationMatrix(R_FM);
 
   q = math::rotmat2rpy(Rproj_FM.matrix());
   return *this;
